Fix GreatDancer race condition on SET_ADDRESS.
The SET_ADDRESS request is only supposed to take affect after its full
transaction (including ACK) is complete, but we were using a non-blocking
ACK request just before setting the actual address.

As a result, in some cases we'd set the address too early, and the ACK
would be lost. This fixes the error by converting the ACK transmission to
blocking, ensuring it's complete before our address change takes effect.