Backend using USBAnywhere vulnerability in Supermicro BMCs
The virtual media functionality in Supermicro X9, X10, and X11 platform
BMCs is implemented such that clients are given raw access to USB
endpoints.  This was discovered along with multiple security
vulnerabilities that have been collectively dubbed USBAnywhere.  The
backend implementation in this commit was created as part of the
research that discovered USBAnywhere.

Additional details on USBAnywhere can be found at
https://github.com/eclypsium/USBAnywhere.  Information on firmware
updates that fix these vulnerabilities can be found at
https://supermicro.com/support/security_BMC_virtual_media.cfm.  Only
unpatched systems can be used as a FaceDancer.