Handle not receiving a short packet if all data has been sent.
Linux seems to not send a zero-length-packet if the data is a multiple
of the maximum packet size for a control transfer, but some devices seem
to be fine with that, and consider the transfer completed if the
specified length of data has been sent. So, now, we also consider a
transfer to be complete if the specified length of data has been sent.