Allow supported_languages = None
This replies to LANGID (string descriptor zero) requests with a stall
which is appropriate for devices that do not implement string
descriptors.

The spec specifically says "USB devices that omit all string descriptors
must not return an array of LANGID codes." Whether that is a stall or an
empty LANGID array appears to be open to interpretation. With this
change we can either stall (with None) or return an empty array (with an
empty tuple).