Fix USBProxy after changes for alternate interface settings.
The active_interfaces attribute on USBConfiguration needs to be set up
when the configuration is activated by a SET_CONFIGURATION request, and
updated on a SET_INTERFACE request.

This was done in our standard request handlers, but not in the proxy.

USBProxyDevice was also maintaining its own lookup table of endpoints,
populated from the configuration. I've removed this because the
configuration has a get_endpoint() method which will do this lookup
instead, taking into account the current interface settings.