test: reset device state before setting interface
Stress and Transfers tests reset the device state (including the expected
endpoint transfer length) only after changing the interface altsetting
(which should restart endpoint transfers with the new length).

This causes a problem in the case when the backend schedules an endpoint
transfer with the old transfer length before the new one gets set.

Reset the device state before changing the interface altsetting.